# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bus Booking System",
        "version": "0.1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "Authenticate User and return bearer Token",
                "operationId": "be2e78e5f93182d4403f839cc516d312",
                "requestBody": {
                    "description": "Pass User data(email , password)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "user@user.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "format": "email",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "if any error happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Ann error occured"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "description": "retrieve authed user profile",
                "operationId": "cd912e022fa4e7d00a92c3b77225aad6",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "If any error happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Not Authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "logout Authenticated User",
                "operationId": "0b366af9c6842623aab9c8827c519b89",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "if any error happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Not Authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "description": "refresh Authenticated User token.",
                "operationId": "1a1e54a850ca4218b0388d92093ab652",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "if any error happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Not Authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}